requirements.txt
python-telegram-bot==20.6
better-profanity==0.7.0
python-dotenv==1.0.0 